% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCellCountsEPIC.r
\name{estimateCellCountsEPIC}
\alias{estimateCellCountsEPIC}
\title{Function to load reference data and process with user supplied EPIC data and do cell composition prediction}
\usage{
estimateCellCountsEPIC(userData, compositeCellType = "Blood",
  processMethod = "auto", probeSelect = "auto", cellTypes = c("CD8T",
  "CD4T", "Bcell", "Mono", "Gran"), returnAll = FALSE, meanPlot = FALSE,
  verbose = TRUE)
}
\arguments{
\item{userData}{is an RGChannelSet}
}
\description{
Function to load reference data and process with user supplied EPIC data and do cell composition prediction
}
\examples{
estimateCellCountsEPIC()
}
